Cleaning the code / Refactoring by "nativizing" `utils.js` (& `index.js`) in raw-handling
