+++ packages/blocks/src/api/raw-handling/index.native.js
@@ -1,48 +1,5 @@
-
-/*
-
-Can't call
-
 import {
 	getPhrasingContentSchema,
 } from './utils';
 
-directly since there are some other variables/imports missing
-*/
-
-/**
- * External dependencies
- */
-import { omit } from 'lodash';
-
-const phrasingContentSchema = {
-	strong: {},
-	em: {},
-	del: {},
-	ins: {},
-	a: { attributes: [ 'href' ] },
-	code: {},
-	abbr: { attributes: [ 'title' ] },
-	sub: {},
-	sup: {},
-	br: {},
-	'#text': {},
-};
-
-// Recursion is needed.
-// Possible: strong > em > strong.
-// Impossible: strong > strong.
-[ 'strong', 'em', 'del', 'ins', 'a', 'code', 'abbr', 'sub', 'sup' ].forEach( ( tag ) => {
-	phrasingContentSchema[ tag ].children = omit( phrasingContentSchema, tag );
-} );
-
-/**
- * Get schema of possible paths for phrasing content.
- *
- * @see https://developer.mozilla.org/en-US/docs/Web/Guide/HTML/Content_categories#Phrasing_content
- *
- * @return {Object} Schema.
- */
-export function getPhrasingContentSchema() {
-	return phrasingContentSchema;
-}
+export { getPhrasingContentSchema };
